+++ src/root.c
@@ -1,9 +1,9 @@
 #include <root.h>
 #include <math.h>
 
-double solve(int a, int b, int c, float* x1, float* x2)
+float solve(float a, float b, float c, float* x1, float* x2)
 {
-	double d;
+	float d;
 
 	d = b * b - 4 * a * c;
 	*x1 = (-b + sqrt(d)) / (2 * a);
